+++ src/plugins/rest0.js
@@ -1,27 +1,35 @@
-export default ({ app, router, Vue }) => {
+export default ({app, router, Vue}) => {
 }
 
 // TODO this is still preliminary
 
 export {
+  getAssetClasses,
   getMissionDefs,
 }
 
-const debug = false
+const debug = true
 
-import { axios } from 'plugins/axios'
+import {axios} from 'plugins/axios'
 
-function getMissionDefs (httpEndpoint) {
+function getAssetClasses(httpEndpoint) {
+  return doGet(httpEndpoint, 'assetclasses')
+}
+
+function getMissionDefs(httpEndpoint) {
+  return doGet(httpEndpoint, 'missiondefs')
+}
+
+function doGet(httpEndpoint, route) {
   return new Promise((resolve, reject) => {
     const method = 'GET'
-    const url = `${httpEndpoint}/missiondefs`
-    if (debug) console.debug('getMissionDefs', 'url=', url)
+    const url = `${httpEndpoint}/${route}`
+    if (debug) console.debug('getGet', 'url=', url)
 
     axios({method, url})
       .then(response => {
         if (debug) console.debug(`${method} ${url}: response=`, response)
-        const missionDefs = response.data.data
-        resolve(missionDefs)
+        resolve(response.data.data)
       })
       .catch(e => {
         reject(e)